Clean up provisioner interface
