+++ src/toil/provisioners/clusterScaler.py
@@ -19,10 +19,12 @@ from collections import deque
 from threading import Lock
 
 from bd2k.util.exceptions import require
+from bd2k.util.retry import retry
 from bd2k.util.threading import ExceptionalThread
 from bd2k.util.throttle import throttle
+from itertools import islice
 
-from toil.batchSystems.abstractBatchSystem import AbstractScalableBatchSystem
+from toil.batchSystems.abstractBatchSystem import AbstractScalableBatchSystem, NodeInfo
 from toil.common import Config
 from toil.provisioners.abstractProvisioner import AbstractProvisioner, Shape
 
@@ -407,8 +409,7 @@ class ScalerThread(ExceptionalThread):
                 if estimatedNodes != self.totalNodes:
                     logger.info('Changing the number of %s from %s to %s.', self.nodeTypeString, self.totalNodes,
                                 estimatedNodes)
-                    self.totalNodes = self.scaler.provisioner.setNodeCount(numNodes=estimatedNodes,
-                                                                           preemptable=self.preemptable)
+                    self.totalNodes = self.setNodeCount(numNodes=estimatedNodes, preemptable=self.preemptable)
                     
                     # If we were scaling up the number of preemptable nodes and failed to meet
                     # our target, we need to update the slack so that non-preemptable nodes will
@@ -426,3 +427,194 @@ class ScalerThread(ExceptionalThread):
                     
         self.scaler.provisioner.shutDown(preemptable=self.preemptable)
         logger.info('Scaler exited normally.')
+
+    def setNodeCount(self, numNodes, preemptable=False, force=False):
+        """
+        Attempt to grow or shrink the number of prepemptable or non-preemptable worker nodes in
+        the cluster to the given value, or as close a value as possible, and, after performing
+        the necessary additions or removals of worker nodes, return the resulting number of
+        preemptable or non-preemptable nodes currently in the cluster.
+
+        :param int numNodes: Desired size of the cluster
+
+        :param bool preemptable: whether the added nodes will be preemptable, i.e. whether they
+               may be removed spontaneously by the underlying platform at any time.
+
+        :param bool force: If False, the provisioner is allowed to deviate from the given number
+               of nodes. For example, when downsizing a cluster, a provisioner might leave nodes
+               running if they have active jobs running on them.
+
+        :rtype: int :return: the number of worker nodes in the cluster after making the necessary
+                adjustments. This value should be, but is not guaranteed to be, close or equal to
+                the `numNodes` argument. It represents the closest possible approximation of the
+                actual cluster size at the time this method returns.
+        """
+        for attempt in retry(predicate=self.scaler.provisioner.retryPredicate):
+            with attempt:
+                workerInstances = self.scaler.provisioner.getWorkersInCluster(preemptable)
+                numCurrentNodes = len(workerInstances)
+                delta = numNodes - numCurrentNodes
+                if delta > 0:
+                    logger.info('Adding %i %s nodes to get to desired cluster size of %i.', delta, 'preemptable' if preemptable else 'non-preemptable', numNodes)
+                    numNodes = numCurrentNodes + self._addNodes(workerInstances,
+                                                                numNodes=delta,
+                                                                preemptable=preemptable)
+                elif delta < 0:
+                    logger.info('Removing %i %s nodes to get to desired cluster size of %i.', -delta, 'preemptable' if preemptable else 'non-preemptable', numNodes)
+                    numNodes = numCurrentNodes - self._removeNodes(workerInstances,
+                                                                   numNodes=-delta,
+                                                                   preemptable=preemptable,
+                                                                   force=force)
+                else:
+                    logger.info('Cluster already at desired size of %i. Nothing to do.', numNodes)
+        return numNodes
+
+    def _addNodes(self, nodes, numNodes, preemptable):
+        return self.scaler.provisioner._addNodes(nodes, numNodes, preemptable)
+
+    def _removeNodes(self, instances, numNodes, preemptable=False, force=False):
+        # If the batch system is scalable, we can use the number of currently running workers on
+        # each node as the primary criterion to select which nodes to terminate.
+        if isinstance(self.scaler.leader.batchSystem, AbstractScalableBatchSystem):
+            # iMap = ip : instance
+            ipMap = {instance.privateIP: instance for instance in instances}
+            def _nodeFilter(executorInfo):
+                return not bool(self._considerNodes([(ipMap.get(executorInfo.nodeAddress), executorInfo.nodeInfo)],
+                                                    preemptable=preemptable))
+            with self.scaler.leader.batchSystem.nodeFiltering(_nodeFilter):
+                # while this context manager is active, the batch system will not launch any
+                # news tasks on nodes that are being considered for termination (as determined by the
+                # _nodeFilter method)
+                nodes = self.getClusterInfo(preemptable)
+                # Join nodes and instances on private IP address.
+                nodes = [(instance, nodes.get(instance.privateIP)) for instance in instances]
+                logger.debug('Nodes considered to terminate: %s', ' '.join(map(str, nodes)))
+                nodesToTerminate = self._considerNodes(nodes, force, preemptable=preemptable)
+                nodesToTerminate = nodesToTerminate[:numNodes]
+                if logger.isEnabledFor(logging.DEBUG):
+                    for instance, nodeInfo in nodesToTerminate:
+                        logger.debug("Instance %s is about to be terminated. Its node info is %r. It "
+                                  "would be billed again in %s minutes.", instance.id, nodeInfo,
+                                  60 * self.scaler.provisioner._remainingBillingInterval(instance))
+                instances = [instance for instance, nodeInfo in nodesToTerminate]
+        else:
+            # Without load info all we can do is sort instances by time left in billing cycle.
+            instances = sorted(instances, key=self.scaler.provisioner._remainingBillingInterval)
+            instances = [instance for instance in islice(instances, numNodes)]
+        logger.info('Terminating %i instance(s).', len(instances))
+        if instances:
+            self.scaler.provisioner._logAndTerminate(instances)
+        return len(instances)
+
+    def _considerNodes(self, nodes, force=False, preemptable=False):
+        # Unless forced, exclude nodes with runnning workers. Note that it is possible for
+        # the batch system to report stale nodes for which the corresponding instance was
+        # terminated already. There can also be instances that the batch system doesn't have
+        # nodes for yet. We'll ignore those, too, unless forced.
+        nodesToTerminate = []
+        for node in nodes:
+            staticNodes = self.scaler.provisioner.getStaticNodes(preemptable)
+            if node.publicIP in staticNodes:
+                # we don't want to automatically terminate any statically
+                # provisioned nodes
+                continue
+            if force:
+                nodesToTerminate.append(node)
+            elif node.nodeInfo is not None and node.nodeInfo.workers < 1:
+                nodesToTerminate.append(node)
+            else:
+                # TODO: fix node info __str__
+                logger.debug('Not terminating instances %s. Node info: %s', node, node.nodeInfo)
+        # Sort nodes by number of workers and time left in billing cycle
+        nodesToTerminate.sort(key=lambda (node): (
+            node.nodeInfo.workers if node.nodeInfo else 1,
+            self.scaler.provisioner._remainingBillingInterval(node)))
+        if not force:
+            # don't terminate nodes that still have > 15% left in their allocated (prepaid) time
+            nodesToTerminate = [nodeTuple for nodeTuple in nodesToTerminate if
+                                self.scaler.provisioner._remainingBillingInterval(nodeTuple[0]) <= 0.15]
+        return nodesToTerminate
+
+    def getClusterInfo(self, preemptable):
+        """
+        Returns a dictionary mapping node identifiers of preemptable or non-preemptable nodes to
+        NodeInfo objects, one for each node.
+
+        This method is the definitive source on nodes in cluster, & is responsible for consolidating
+        cluster state between the provisioner & batch system.
+
+        :param bool preemptable: If True (False) only (non-)preemptable nodes will be returned.
+               If None, all nodes will be returned.
+
+        :rtype: dict[str,NodeInfo]
+        """
+        def _getInfo(allMesosNodes, ip):
+            info = None
+            try:
+                info = allMesosNodes[ip]
+            except KeyError:
+                # never seen by mesos - 1 of 3 possibilities:
+                # 1) node is still launching mesos & will come online soon
+                # 2) no jobs have been assigned to this worker. This means the executor was never
+                #    launched, so we don't even get an executorInfo back indicating 0 workers running
+                # 3) mesos crashed before launching, worker will never come online
+                # In all 3 situations it's safe to fake executor info with 0 workers, since in all
+                # cases there are no workers running. We also won't waste any money in cases 1/2 since
+                # we will still wait for the end of the node's billing cycle for the actual
+                # termination.
+                info = NodeInfo(coresTotal=1, coresUsed=0, requestedCores=0,
+                                memoryTotal=1, memoryUsed=0, requestedMemory=0,
+                                workers=0)
+            else:
+                # Node was tracked but we haven't seen this in the last 10 minutes
+                inUse = self.scaler.leader.batchSystem.nodeInUse(ip)
+                if not inUse:
+                    # The node hasn't reported in the last 10 minutes & last we know
+                    # there weren't any tasks running. We will fake executorInfo with no
+                    # worker to reflect this, since otherwise this node will never
+                    # be considered for termination
+                    info.workers = 0
+                else:
+                    pass
+                    # despite the node not reporting to mesos jobs may still be running
+                    # so we can't terminate the node
+            return info
+
+        allMesosNodes = self.scaler.leader.batchSystem.getNodes(preemptable, timeout=None)
+        recentMesosNodes = self.scaler.leader.batchSystem.getNodes(preemptable)
+        provisionerNodes = self.scaler.provisioner._getProvisionedNodes(preemptable)
+
+        if len(recentMesosNodes) != len(provisionerNodes):
+            logger.debug("Consolidating state between mesos and provisioner")
+            ipToInfo = {}
+            # fixme: what happens if awsFilterImpairedNodes is used?
+            # if this assertion is false it means that user-managed nodes are being
+            # used that are outside the provisioners control
+            # this would violate many basic assumptions in autoscaling so it currently not allowed
+            for node, ip in ((node, node.privateIP) for node in provisionerNodes):
+                info = None
+                logger.debug("Worker node at %s is not reporting executor information")
+                if ip not in recentMesosNodes:
+                    # we don't have up to date information about the node
+                    info = _getInfo(allMesosNodes, ip)
+                else:
+                    # mesos knows about the ip & we have up to date information - easy!
+                    info = recentMesosNodes[ip]
+                # add info to dict to return
+                ipToInfo[ip] = info
+            return ipToInfo
+        else:
+            return recentMesosNodes
+
+    def shutDown(self, preemptable):
+        if not self.stop:
+            # only shutdown the stats threads once
+            self.scaler.provisioner._shutDownStats()
+        logger.debug('Forcing provisioner to reduce cluster size to zero.')
+        totalNodes = self.setNodeCount(numNodes=0, preemptable=preemptable, force=True)
+        if totalNodes > len(self.scaler.provisioner.getStaticNodes(preemptable)):  # ignore static nodes
+            raise RuntimeError('Provisioner could not terminate all autoscaled nodes. There are '
+                               '%s nodes left in the cluster, %s of which were statically provisioned' % (totalNodes, len(self.getStaticNodes(preemptable)))
+                               )
+        elif totalNodes < len(self.scaler.provisioner.getStaticNodes(preemptable)):  # ignore static nodes
+            raise RuntimeError('Provisioner incorrectly terminated statically provisioned nodes.')
